requirements.txt
Django==3.0.8
django_filter==2.3.0
Pillow==7.2.0
django-bootstrap-form